requirements.txt
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
googlemaps==4.10.0
rich==13.8.1
gsheet-pandas==0.2.8
alive-progress==3.1.5